1 Вынес интерфейс файловой системы в отдельный файл в api 2 Импортировал интерфейсы в store в client
