+++ api/src/services/app.fileSystemService.ts
@@ -1,22 +1,9 @@
 import { Injectable } from "@nestjs/common";
+import {FileModel,FolderModel} from "../model/FileSystemModel";
 
 const fs = require("fs");
 const path = require("path");
 const childProcess = require('child_process');
-enum EtypeItem {
-  file = "file",
-  folder = "folder",
-}
-
-interface BaseModel<T extends EtypeItem> {
-  type: T;
-  name: string;
-  lastDateChange: Date;
-}
-export interface FileModel extends BaseModel<EtypeItem.file> {
-  size: number;
-}
-export interface FolderModel extends BaseModel<EtypeItem.folder> {}
 
 @Injectable()
 export class AppFileSystemService {
