+++ client/src/main.ts
@@ -1,7 +1,8 @@
 import { createApp } from 'vue'
 import App from './App.vue'
 import store from './store'
-
+import {FileModel,FolderModel} from '../../api/src/model/FileSystemModel'
+export {FileModel,FolderModel};
 
 const app = createApp(App)
 // app.provide("store", store)
